fix: apply valid PR #56 review suggestions
- Use cardinality() instead of array_length() to reliably reject empty
  category arrays (array_length returns NULL for empty arrays, bypassing CHECK)
- Move SerializedPreference and serializePreference to preference-serializer.ts
  for consistency with session-serializer.ts pattern
- Fix misleading RLS comment that referenced sessions migration
